Fix missing implied edges in the transitive closure
When creating edges for nodes "out of order" (e.g. not creating the entire graph top down), some edges wouldn't be created for nodes that could be reached indirectly.

For example, when connecting nodes in the following order:

(5) -> (8)
(5) -> (9)
(1) -> (3)
(3) -> (5)

there would be no implied edges created to represent the indirect connection between nodes 1 & 8 and 1 & 9